separated utilities and rootchain helpers
